@@ -1,11 +1,11 @@
 name: flutter_pattern_locker
 description: PatternLocker
-version: 0.0.7
+version: 1.0.0
 homepage: https://github.com/Leekajun/flutter_pattern_locker
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
