@@ -21,9 +21,9 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
   final String title;
-  LockConfig config = LockConfig();
+
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
@@ -32,7 +32,7 @@ class _MyHomePageState extends State<MyHomePage> {
   String tip = '请设置密码';
   String subTip = '';
   LockConfig config = LockConfig();
-  String pwd;
+  String? pwd;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
